Move APIClient tests into single file
For some reason this was spread across two files.
